refactor(core): re-org files
