+++ packages/core/src/routes/consts.ts
@@ -1,0 +1,8 @@
+import { signIn } from '@/env/consts';
+
+export const routes = Object.freeze({
+  signIn: {
+    credentials: signIn,
+    consent: signIn + '/consent',
+  },
+});
